Force relationships to use the proxy when getting primary key
Fixes #18

When the relationships (HasMany, HasOne, BelongsTo) retrieve a primary key in their `getPrimaryKeyValue` methods, they do so with code like this:

```
(entity as any)[this.primaryKey]
```

It turns out the `entity` in this case is not the proxy, but is instead the "raw" underlying object.  This is because the getters for relationships are functions that are added to the prototype of the underlying Entity class, rather than being intercepted somehow by the proxy.  As a result, the "this" for that getter is the underlying Entity instance, not the proxy.

The fix is simple enough - just change the above to:

```
(entity.currentEntity as any)[this.primaryKey]
```