Include entities only once when exporting
For Entity classes that are subclasses of other Entity classes, the Entity instances will appear in the "byId" index of its own Entity class and all its super classes.  This is intentional, but when exporting entities it can lead to entities appearing multiple times in the export.

This PR prevents this from happening by making sure an entity will only be included if the export is currently exporting the byId for that entity's actual class.